+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix text type to show at `assembly-option-item-children-description`.
+
+### Changed
+- Save item group type in state.
 
 ## [2.1.0] - 2019-07-26
 ### Added